Added per model gamma correction and all characters now have a white point light below them. 💡

+++ source/Skynet/Pathfinding.cpp
@@ -144,7 +144,7 @@ bool IsCellValid(int x, int z)
 
 float CalculateEuclideanDistance(int x, int z, Node goal)
 {
-	float dist = (sqrt((x - goal.x) * (x - goal.x)+ (z - goal.z) * (z - goal.z)));
+	float dist = (sqrtf((x - goal.x) * (x - goal.x)+ (z - goal.z) * (z - goal.z)));
 	return dist;
 }
 
@@ -248,7 +248,7 @@ ML_Vector<Node> CalculateAStarPath(PathfindingMap gridValues, TransformComponent
 		GridPosition currPos; //= *openList.begin();
 		float cheapest = FLT_MAX;
 		int index = 0;
-		for (int i = 0; i < openList.size(); i++)
+		for (unsigned i = 0; i < openList.size(); i++)
 		{
 			if (nodeMap[openList[i].x][openList[i].z].f < cheapest)
 			{
